requirements.txt
appdirs==1.4.3
asgi-redis==1.4.3
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.11.2
Automat==0.7.0
beautifulsoup4==4.6.3
boto3==1.9.50
botocore==1.12.50
bs4==0.0.1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.4.2
cssselect==1.0.3
defusedxml==0.5.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.3
django-crispy-forms==1.7.2
django-filter==2.0.0
django-heroku==0.3.1
django-localflavor==2.1
django-qr-code==1.0.0
django-random-id-generator==0.1.0
django-storages==1.7.1
djangorestframework==3.9.0
docutils==0.14
fake-useragent==0.1.11
gunicorn==19.9.0
hiredis==0.3.0
html5lib==1.0.1
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
jmespath==0.9.3
lxml==4.2.5
Markdown==3.0.1
msgpack==0.5.6
msgpack-python==0.5.6
numpy==1.15.4
oauthlib==2.1.0
opencv-python==3.4.3.18
parse==1.9.0
Pillow==5.3.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pycparser==2.19
pyee==5.0.0
PyHamcrest==1.9.0
PyJWT==1.6.4
PyPDF2==1.26.0
pyppeteer==0.0.25
pyquery==1.4.0
PySocks==1.6.8
pystache==0.5.4
python-dateutil==2.7.5
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.7
qrcode==6.0
redis==3.0.1
reportlab==3.5.11
requests==2.20.1
requests-html==0.9.0
requests-oauthlib==1.0.0
s3transfer==0.1.13
selenium==3.141.0
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
static==1.1.1
static3==0.7.0
tqdm==4.28.1
twilio==6.20.0
txaio==18.8.1
urllib3==1.24.1
w3lib==1.19.0
webencodings==0.5.1
websockets==7.0
whitenoise==4.1.2
xmltodict==0.11.0
zope.interface==4.6.0
